Middle of refactor - started seperating to handlers per resource type
